refactor: domain parts from DotChaser into separate project
