+++ DotChaser/DotChaser.Client/DotChaser.Client.csproj
@@ -7,4 +7,8 @@
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\DotChaser\DotChaser.csproj" />
+    </ItemGroup>
+
 </Project>
