+++ HiveShard.sln
@@ -51,6 +51,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DotChaser", "DotChaser", "{
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotChaser.Client", "DotChaser\DotChaser.Client\DotChaser.Client.csproj", "{962DB4FC-B650-43A9-87E5-E657F7BC1B1E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotChaser", "DotChaser\DotChaser\DotChaser.csproj", "{F599ED63-E969-46BA-8998-9BA4BA0B8D79}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -136,6 +138,10 @@ Global
 		{962DB4FC-B650-43A9-87E5-E657F7BC1B1E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{962DB4FC-B650-43A9-87E5-E657F7BC1B1E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{962DB4FC-B650-43A9-87E5-E657F7BC1B1E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F599ED63-E969-46BA-8998-9BA4BA0B8D79}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F599ED63-E969-46BA-8998-9BA4BA0B8D79}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F599ED63-E969-46BA-8998-9BA4BA0B8D79}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F599ED63-E969-46BA-8998-9BA4BA0B8D79}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{8A33580F-899C-4BBE-9E0F-029E5ECF127F} = {DC5DCDF9-C8CB-4A8F-8D1A-ACE90F3EFAA9}
@@ -151,5 +157,6 @@ Global
 		{017BF16D-72E7-421D-998D-F5BD66B0FF28} = {0C8A359F-A14A-48B5-9EF3-9C297325C294}
 		{97B5A89D-FBCA-40F4-940F-5AB0A3A4AAEF} = {0C8A359F-A14A-48B5-9EF3-9C297325C294}
 		{962DB4FC-B650-43A9-87E5-E657F7BC1B1E} = {D72BAF2E-D112-4E25-9CE7-2BE8D956750B}
+		{F599ED63-E969-46BA-8998-9BA4BA0B8D79} = {D72BAF2E-D112-4E25-9CE7-2BE8D956750B}
 	EndGlobalSection
 EndGlobal
